added new controller
